\name{GEOsymbols}
\alias{GEOsymbols}

\title{GEOsymbols}
\description{
  Plot a set of Geological Symbols
}
\usage{
GEOsymbols()
}

\details{
  Currently the choices in symbols are:
\begin{verbatim}
contact
anticline
syncline
OverTurned-ant
OverTurned-syn
perp
thrust
normal
dextral
sinestral
detachment
bcars
\end{verbatim}


  }
\value{
 Graphical Side effect
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{bcars, thrust, teeth, SynAnticline, SSfault, horseshoe,
  strikeslip, OverTurned, normalfault, PointsAlong}
\examples{


GEOsymbols()


}

\keyword{aplot}
