\name{lcgc}
\alias{lcgc}

\title{local coordinates to Global}
\description{
  OLD projection sometimes used in Lees' tomography. No need for
  projection data, it is included in the code.
}
\usage{
lcgc(phiorg, lamorg, ex, why)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{phiorg}{ lat origin}
  \item{lamorg}{lon origin}
  \item{ex}{coordinate, km }
  \item{why}{coordinate, km  }
}
\details{
 This may be defunct now.
}
\value{
   \item{phi}{lat}
  \item{lam}{lon}
  
}
\note{Orignally from R. S. Crosson}
\author{Jonathan M. Lees<jonathan.lees.edu>}


\seealso{gclc}
\examples{

}

\keyword{misc}
