\name{ll2xyz}
\Rdversion{1.1}
\alias{ll2xyz}
\title{LAT-LON to xyz
}
\description{LAT-LON to xyz
}
\usage{
ll2xyz(lat, lon)
}
\arguments{
  \item{lat}{latitude
}
  \item{lon}{longitude
}
}

\value{3-vector}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{Lll2xyz, Lxyz2ll, xyz2ll
}
\examples{

ll2xyz(12, 289)

}
\keyword{misc}
