\name{Lll2xyz}
\Rdversion{1.1}
\alias{Lll2xyz}
\title{List Lat-Lon to cartesian XYZ
}
\description{List Lat-Lon to cartesian XYZ
}
\usage{
Lll2xyz(lat, lon)
}
\arguments{
  \item{lat}{latitude
}
  \item{lon}{longitude
}
}

\value{list(x,y,z)
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{ll2xyz, Lxyz2ll, xyz2ll
}
\examples{

Lll2xyz(23, 157)



}
\keyword{misc}
