\name{Lxyz2ll}
\Rdversion{1.1}
\alias{Lxyz2ll}
\title{Cartesian to Lat-Lon
}
\description{Cartesian vector to Lat-Lon List
}
\usage{
Lxyz2ll(X)
}
\arguments{
  \item{X}{list, x,y,z}
}

\value{list of lat and lon
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{xyz2ll}
\examples{

Lll2xyz(23, 157)

}
\keyword{misc}
