\name{MAPconstants}
\alias{MAPconstants}

\title{Set Various Map Constants}
\description{
 Used for retrieval when doing projections
}
\usage{
MAPconstants()
}

\details{
  These include a sime list of:
  DEG2RAD, RAD2DEG, A.MAPK, E2.MAPK, E2.GRS80, E.MAPK, E1.MAPK, TwoE.MAPK, R.MAPK, FEET2M, M2FEET

}
\value{
  List of constants for Projections
}
\references{
Snyder, John P., Map Projections- a working manual, USGS, Professional Paper, 1987.
}
\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{XY.GLOB, projtype, utm.sphr.xy}
\examples{
MAPconstants()


}

\keyword{misc}
