% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plain}
\alias{plain}
\title{style map}
\usage{
plain(point = list(colour = "black", size = 1, symbol = 1),
  poly = list(colour = "white", pattern = 1), line = list(colour = "black",
  width = 1, pattern = 1))
}
\description{
function mapping features to style
}

