\name{printGEOmap}
\alias{printGEOmap}
\title{printGEOmap }
\description{Print information on GEOmap strokes
}
\usage{
printGEOmap(G)
}
\arguments{
  \item{G}{GEOmap }
}
\details{
  Prints the full STROES list as a dataframe.
}
\value{
  Side Effects 
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{printGEOinfo}
\examples{

data(coastmap)
printGEOmap(coastmap)

}
\keyword{misc}
