\name{rotdelta4}
\Rdversion{1.1}
\alias{rotdelta4}
\title{rotation about Z-axis}

\description{rotation about Z-axis
}

\usage{
rotdelta4(delta)
}
\arguments{
  \item{delta}{angle in degrees}
}

\value{Matrix for rotation
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{roty4, rotx4, trans4}
\examples{
rotdelta4(23)


}
\keyword{misc}
