\name{trans4}
\Rdversion{1.1}
\alias{trans4}
\title{Translation matrix
}
\description{Translation matrix for rotations
}
\usage{
trans4(vec)
}
\arguments{
  \item{vec}{ 3 vector}
}

\value{ 4 by 4 matrix
}
\references{Rogers and Adams
}
\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{rotx4, roty4, rotdelta4
}
\examples{
trans4(c(0,0,0))

}
\keyword{misc}
