\name{X.prod}
\Rdversion{1.1}
\alias{X.prod}
\title{Cross Product
}
\description{Vector Cross Product for spatial cartesian vectors
}
\usage{
X.prod(a, b)
}
\arguments{
  \item{a}{3-vector
}
  \item{b}{3-vector
}
}

\value{ 3-vector
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\examples{

v1 = c(1,1,1)
v2= c(-1, -1, 1)
X.prod(v1, v2)


}
\keyword{misc}
